Abstract Board and Field from Game
Field already existed, but was coupled with Game. Make it fully
abstract, moving the game-specific logic to Game and making it an
observer of individual Fields.

Board is a newly-introduced class for keeping a collection of fields.

Generation of random board is now done in a newly-introduced module
gen.js. No CSP smarts yet.