index: Remove delay before click event is issued
Browsers impose a delay before click event is issued to find out whether
what the user wants is actually a double click. That delay however makes
the game appear laggy.

The event that is issued right away is touchstart resp. touchend, but
then one loses an easy support for long touch. Reimplementing that is
not entirely trivial: one needs a timer and then needs to replicate all
the logic that the browser already does to figure out if there was a
motion, how far the finger traveled, whether it landed on the same
entity, and how long does "long" take. Let's not have this either.

The "viewport" meta tag basically tells the browser that the web page is
mobile-ready, that it shouldn't render to a larger viewport and then
pan over it. Browsers also interpret this as that there's no need for
double-touch support, and suppress the delay.

Besides this, there's also "touch-action: manipulation", a CSS tag that
supposedly disables double-click-to-zoom on certain elements, therefore
allowing a browser to again suppress the delay. But Firefox doesn't
appear to support this despite having a config key to enable this
support. It does work on Chrome, but so does the viewport meta tag.