Use bloom filters to check whether to deduceCoupled
Even though there is some work on RB rebalancing, this is not a full
solution. But we enable bloom filering just to get a feel for how it
currently stands, and there is actually a slight degradation of
performance. Even though bloom filter saves about two thirds of
deduceCoupled calls, the bloom filtering itself adds enough overhead
back that it's slightly worse overall.