asm: Further split conses into groups
Instead of having one store for all conses, with several indices, split
the storage into groups as well. Thus in order to deduceCoupled vs. that
one group, it's not necessary to go through the indirect indexing of
order[g], but it is possible to simply walk the group as such.

Together with the previous splitting change, this delivers about 20 %
performance boost in the newly-added deterministic gen-test.