asm.js: Move substKnowns below deduction
The current placemen of substKnowns is between deduceSimple and
deduceCoupled. Both of these contribute new knowns that substKnowns then
substitutes to constraints. The placement is illogical--it should either
first substitute knowns, and then do all deductions, or vice versa. Thus
move the substKnowns block to the bottom.

This improves the deterministic gen-test performance by almost 10%.