Allow more nuanced autoplay selection
Instead of passing ignoreCoupled to the solver, pass a more granular
strength, with varying levels of strength corresponding to different
decisions that the solver is allowed to do.

To support a mode where only primary constraints are solved, add flags
to each constraint. This allows definition of strengthPrimaryZeroes,
which describes exactly the usual mode where zeroes on board are
auto-opened, but nothing else is done. strengthZeroes is then a similar
mode that allows solving also deduced constraints with zero RHS, which
leads to an interesting autoplay mode where fields open if they have as
many flags as their number indicates (but flags are never auto-placed).